whatpm/t/tokenizer-test-2.dat

#data escaped
<!DOCTYPE html><html><head><title></title></head><body><p>\u0000</p></body>
#errors
NULL
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>
|       "\uFFFD"
